+++ data/readme.md
@@ -1,0 +1,14 @@
+## Info on datasets
+
+| **Batch** | **Cell Type**                              | **Perturbation** | **Control** | **Normalization**     |
+|-----------|---------------------------------------------|------------------|-------------|------------------------|
+| Batch1    | Myoblasts - C2C12                                    | Selenon-KO       | WT cells    | Negcon (WT cells)      |
+| Batch2    | Primary fibroblasts                         | Selenon-KO       | WT cells    | Negcon (WT cells)      |
+| Batch3    | Primary fibroblasts  <br> (Diaphragm and Gastrocnemius) | Selenon-KO       | WT cells    | Negcon (WT cells)      |
+| Batch4    | Primary myoblasts  <br> (Quadriceps and Gastrocnemius) | Selenon-KO       | WT cells    | Negcon (WT cells)      |
+| Batch5    | Myoblasts - C2C12 | Selenon-KO       | WT cells    | Negcon (WT cells)      |
+
+### Pickle files
+
+- `batch*.pkl` - Batch data transposed and label column added
+- `b1b5.pkl` - combines Batch1 and Batch5 data
